requirements.txt
# Core FastAPI e uvicorn
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Banco de dados
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# OCR e Visão Computacional
paddleocr==2.7.3
paddlepaddle>=2.4.0
opencv-python==4.6.0.66
Pillow==10.1.0
pytesseract==0.3.10

# Transformers para TrOCR
transformers==4.35.2
torch>=2.0.0
torchvision>=0.15.0
accelerate==0.25.0

# NLP
spacy==3.7.2
langchain==0.0.339
langchain-community==0.0.3

# Embeddings e Busca Vetorial
sentence-transformers==2.7.0
huggingface-hub==0.19.4
qdrant-client==1.7.0
faiss-cpu==1.7.4

# Utilitários
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.3
pandas==2.1.3
uuid7==0.1.0

# LLM providers
groq==0.30.0
openai==1.3.7

# Logging e monitoramento
loguru==0.7.2

# Validação e utilitários
validators==0.22.0
python-magic==0.4.27

# Interface web (opcional)
streamlit==1.29.0 
 